requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-Login==0.6.2
python-dotenv==1.0.1
Werkzeug==2.3.7
pymongo==4.6.1
